requirements.txt
numpy>=1.26.0
pillow>=10.2.0
moviepy>=1.0.3
streamlit>=1.32.0
oemer==0.1.5
onnx>=1.15.0
onnxruntime>=1.17.0
tensorflow>=2.15.0
keras>=2.15.0
pandas>=2.2.0